Se arreglaron los paths de los sing ups, ademas de acomodar las pantallas en sus respetivas carpetas, ademas de limpiar el codigo y se agregaron dos widgets en singup4
