+++ lib/screens/signup4/signup4.dart
@@ -1,129 +1,37 @@
 import 'package:auto_matic/config/config.dart';
+import 'package:auto_matic/screens/signup4/widgets/accesswithsocial.dart';
+import 'package:auto_matic/screens/signup4/widgets/createaccountsection.dart';
 
-/// Sección de creación de cuenta
-class CreateAccountSection extends StatelessWidget {
+class SinguUp4 extends StatelessWidget {
+  const SinguUp4({super.key});
+
   @override
   Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Text('Crear cuenta', style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold, color: Config.secondColor)),
-        SizedBox(height: 20),
-        Container(
-          decoration: BoxDecoration(
-            color: Config.firstColor,
-            borderRadius: BorderRadius.circular(10),
-            boxShadow: [
-              BoxShadow(
-                color: Colors.grey.withOpacity(0.5),
-                spreadRadius: 2,
-                blurRadius: 5,
-                offset: Offset(0, 3),
-              ),
-            ],
-          ),
-          padding: EdgeInsets.all(20),
-          child: Column(
-            children: [
-              Row(
-                children: [
-                  Expanded(
-                    child: TextFormField(
-                      decoration: InputDecoration(
-                        labelText: 'Nombre',
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(10),
-                        ),
-                      ),
-                    ),
-                  ),
-                  SizedBox(width: 20),
-                  Expanded(
-                    child: TextFormField(
-                      decoration: InputDecoration(
-                        labelText: 'Apellidos',
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(10),
-                        ),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-              SizedBox(height: 20),
-              TextFormField(
-                decoration: InputDecoration(
-                  labelText: 'Correo Electrónico',
-                  border: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(10),
-                  ),
-                ),
-              ),
-              SizedBox(height: 20),
-              TextFormField(
-                obscureText: true,
-                decoration: InputDecoration(
-                  labelText: 'Contraseña',
-                  border: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(10),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-        SizedBox(height: 20),
-        ElevatedButton(
-          onPressed: () {
-            // Tu lógica aquí
-          },
-          style: ElevatedButton.styleFrom(
-            backgroundColor: Config.secondColor, // Cambia el valor hexadecimal aquí
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(10),
-            ),
-          ),
-          child: Text(
-            'Continuar',
-            style: TextStyle(
-              color: Config.firstColor, // Cambia el valor hexadecimal aquí
-              fontSize: 18,
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-// Sección de acceso con redes sociales
-class AccessWithSocialMediaSection extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Text('Accede con', style: TextStyle(fontSize: 24,fontWeight: FontWeight.bold, color: Config.secondColor)),
-        SizedBox(height: 10),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.center,
+    return Container(
+      color: Config.firstColor, // Fondo blanco
+      padding: EdgeInsets.all(16.0),
+      child: Column(
+          mainAxisAlignment: MainAxisAlignment.spaceAround,
           children: [
-            Container(
-              width: 60,
-              height: 60,
-              color: Colors.blue,
-              child: Center(child: Text('Redes Sociales')),
+            // Sección de creación de cuenta
+            Expanded(
+              flex: 2,
+              child: CreateAccountSection(),
             ),
-            SizedBox(width: 20),
-            Container(
-              width: 60,
-              height: 60,
-              color: Colors.green,
-              child: Center(child: Text('Google')),
+            SizedBox(width: 40),
+            // Sección de acceso con redes sociales
+            Expanded(
+              flex: 1,
+              child: AccessWithSocialMediaSection(),
             ),
-          ],
-        ),
-      ],
+            Expanded(
+              flex: 2,
+              child: TimelineWidget(currentPage: 4),
+            ),
+            SizedBox(width: 40),
+            //ConstructorBody()
+          ]
+      ),
     );
   }
 }
-
